fix: refactor cloud accounts page
